fetch movie list inside provider
